save month to localStorage
