+++ src/components/ImportantCritter/ImportantCritter.js
@@ -5,14 +5,15 @@
  *  Copyright (c) 2020 Lucy Tan
  */
 
-import React, { useContext, useState } from "react";
+import React, { useContext } from "react";
 import { Button } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 
 import { CritterTable } from "../CritterTable/CritterTable";
 import { DropdownSelector } from "../DropdownSelector/DropdownSelector";
-import { getCrittersAvailableByMonth, getCrittersLeavingByMonth, getQueryParam, monthNameToNumMap } from "../../data/utils";
+import { getCrittersAvailableByMonth, getCrittersLeavingByMonth, monthNameToNumMap } from "../../data/utils";
 import { CritterDataContext } from "../../context/CritterDataContext";
+import LocalStorageContext from "../../context/LocalStorageContext";
 
 const useStyles = makeStyles((theme) => ({
   buttonContainer: {
@@ -43,8 +44,8 @@ const useStyles = makeStyles((theme) => ({
 const monthNames = Object.keys(monthNameToNumMap)
 
 export const ImportantCritterSection = ({ critter }) => {
-  const { fishWithDates, insectWithDates } = useContext(CritterDataContext)
-  const [selectedMonth, setSelectedMonth] = useState(getMonth());
+  const { fishWithDates, insectWithDates } = useContext(CritterDataContext);
+  const { selectedMonth, setSelectedMonth } = useContext(LocalStorageContext);
   const critterWithDates = critter === 'fish' ? fishWithDates : insectWithDates;
   const classes = useStyles();
 
@@ -81,16 +82,4 @@ export const ImportantCritterSection = ({ critter }) => {
       />
     </>
   );
-};
-
-/**
- * Returns the current month or the month specified in queryParams
- */
-const getMonth = () => {
-  const queryMonth = getQueryParam().month;
-  const isQueryMonthInvalid = isNaN(monthNameToNumMap[queryMonth]);
-  const date = new Date();
-  const monthName = isQueryMonthInvalid ? date.toLocaleDateString("default", { month: "long" }) : queryMonth;
-
-  return monthName
 };
